Fixed Camera Movement
Saw the camera still not turning properly if tilted. What I did here was
tilt the camera back up/down to perfectly horizontal before panning it
left or right, then I tilted it back to the way it was. It seems
quaternions aren't quite as perfect as I thought before. Let me know if
you spot a bug in this or find out about an easier way.