CollisionClass
Made a skeleton of a collision class. I'm not sure if we should make an
instance of this class in the world, or if we should just call its
members statically. I also made the CheckLocation method inside the
levelclass static, so that the collision class didn't need to have an
instance of the level (or so we didn't need to pass around a level
object).