Collision Change
I realized immediately after that last commit that we'll probably have
instances of different levels, which means that the CheckLocation
function should not be static. So I changed it back, and I also changed
the CheckCollision parameters to take a level pointer. I think it makes
a bit more sense this way.