Mouse input
Sorry for being a day late. In any case, mouse input has been added. I
also fixed the mouse to the center of the screen and made it invisible.
If you want that changed just look to the InitializeWindows() function
and the Frame() function. I'm going to clean up the cinecamera class a
bit also.