Updates to LightMask
I now have a public static LightState for the light mask. I have it set
in the frame() function of lightmask.cpp so that when the state is set
to LightMask::ON the light is set to 0.0f and when it's LightMask::OFF
the light is set to 0.8f. Having it static makes it easy to use from
anywhere in my opinion, but feel free to change it or have me change it.
Also the values are easy to mess around with.