Duplicate Level Fixed
There seems to be a lot of changes since the last commit, but I assure
you, it's not detrimental. It looks like 2 things I did which have no
effect are implement a shutdown in the DoorObject and WallObject
classes, and uncomment the code in the gameobject shutdown. I don't
think those will cause any trouble, but still keep them in mind.