Lot's of stuff... namely, AI patrol paths
I did a bunch of stuff that I can't quite remember, but I'm committing
this now since it's working. Going to add one last thing to the patrol
path feature, and since works pretty funny, you may want to just let me
handle it for when I go to put together the level (you have to make the
path manually for each enemy).