Level Build **READ DESCRIPTION**
I'm not sure about your machine, but on my machine, there seem to be
some hardware difficulties or something. What I mean is that, if I make
the level too big, it won't load, it will give me the Could not load
model pop up and just break. I want to text this out on a better
machine, but for now, I've build the level on a 12x12 grid, and it runs
fine. I wanted to go 50x50, but that died, so I dropped it to 25x25 and
that died at runtime (weird eh?). So now I'm going to commit here, then
move to my bigger laptop and see if it really is a hardware issue.
Things to do with this build are bind everything to their specific
rooms. I think I can do that easy enough, but let me test this thing out
on my other machine first.