Little "Locked Door" Sound
Just giving the player a bit more indication that they can't open a door
(though the sound effect for it kind of sucks actually). You wouldn't
believe how hard it was looking for a sound for this. I had to settle
for the sound of a purse snapping shut... Close enough, I thought.